Get rid of as many compiler warnings as possible.
Add missing offbits seek when reading BMP files.

Add -OPT:Olimit=4000 to SGI compiler options (for keyboard_ui)


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@2253 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_PNG_Image.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_PNG_Image.cxx,v 1.1.2.5 2002/04/24 18:29:05 easysw Exp $"
+// "$Id: Fl_PNG_Image.cxx,v 1.1.2.6 2002/05/24 14:19:19 easysw Exp $"
 //
 // Fl_PNG_Image routines.
 //
@@ -89,8 +89,8 @@ Fl_PNG_Image::Fl_PNG_Image(const char *png) // I - File to read
   if ((info->color_type & PNG_COLOR_MASK_ALPHA) || info->num_trans)
     channels ++;
 
-  w(info->width);
-  h(info->height);
+  w((int)(info->width));
+  h((int)(info->height));
   d(channels);
 
   if (info->bit_depth < 8)
@@ -135,5 +135,5 @@ Fl_PNG_Image::Fl_PNG_Image(const char *png) // I - File to read
 
 
 //
-// End of "$Id: Fl_PNG_Image.cxx,v 1.1.2.5 2002/04/24 18:29:05 easysw Exp $".
+// End of "$Id: Fl_PNG_Image.cxx,v 1.1.2.6 2002/05/24 14:19:19 easysw Exp $".
 //
